requirements.txt
python-dotenv
langchain
langchain-community
openai
PyPDF2
faiss-cpu
streamlit
tiktoken
-e .